# openapi.yml
openapi: 3.0.3
info:
  title: Celebrity Strike API
  description: API for managing games, images, and game items in the Celebrity Strike application
  version: 1.0.0
  contact:
    name: Celebrity Strike Team
servers:
  - url: /api/v1
    description: API base path
paths:
  /games:
    get:
      summary: Get all games
      description: Retrieves a list of all games
      operationId: getGames
      tags:
        - Games
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListGameRespDto'
    post:
      summary: Create a new game
      description: Creates a new game with associated images and items
      operationId: saveGames
      tags:
        - Games
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameReqDto'
      responses:
        '200':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
  /games/id/{id}:
    get:
      summary: Get game by ID
      description: Retrieves a game by its ID
      operationId: getGamesById
      tags:
        - Games
      parameters:
        - name: id
          in: path
          description: Game ID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
  /games/{name}:
    get:
      summary: Get game by name
      description: Retrieves a game by its name
      operationId: getGamesByName
      tags:
        - Games
      parameters:
        - name: name
          in: path
          description: Game name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
    put:
      summary: Update game by name
      description: Updates an existing game's fields by its name
      operationId: updateGame
      tags:
        - Games
      parameters:
        - name: name
          in: path
          description: Game name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameUpdateReqDto'
      responses:
        '200':
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameRespDto'
    delete:
      summary: Delete game by name
      description: Deletes a game and all its associated images and items
      operationId: deleteGame
      tags:
        - Games
      parameters:
        - name: name
          in: path
          description: Game name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

  /images:
    post:
      summary: Create a new image
      description: Creates a new image for a game
      operationId: createImage
      tags:
        - Images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreateReqDto'
      responses:
        '200':
          description: Image created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultImageRespDto'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultImageRespDto'

  /images/{id}:
    put:
      summary: Update image
      description: Updates an existing image
      operationId: updateImage
      tags:
        - Images
      parameters:
        - name: id
          in: path
          description: Image ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageUpdateReqDto'
      responses:
        '200':
          description: Image updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultImageRespDto'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultImageRespDto'
    delete:
      summary: Delete image
      description: Deletes an image
      operationId: deleteImage
      tags:
        - Images
      parameters:
        - name: id
          in: path
          description: Image ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Image deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

  /game-items:
    post:
      summary: Create a new game item
      description: Creates a new game item for a game
      operationId: createGameItem
      tags:
        - Game Items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameItemCreateReqDto'
      responses:
        '200':
          description: Game item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameItemRespDto'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameItemRespDto'

  /game-items/{id}:
    put:
      summary: Update game item
      description: Updates an existing game item
      operationId: updateGameItem
      tags:
        - Game Items
      parameters:
        - name: id
          in: path
          description: Game Item ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameItemUpdateReqDto'
      responses:
        '200':
          description: Game item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameItemRespDto'
        '404':
          description: Game item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultGameItemRespDto'
    delete:
      summary: Delete game item
      description: Deletes a game item
      operationId: deleteGameItem
      tags:
        - Game Items
      parameters:
        - name: id
          in: path
          description: Game Item ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Game item deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
        '404':
          description: Game item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

  /chat/send-message/{id}:
    post:
      summary: Send message to chat
      description: Sends a message to a specific chat by ID
      operationId: sendMessage
      tags:
        - Chat
      parameters:
        - name: id
          in: path
          description: Chat ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageToChatReq'
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

components:
  schemas:
    ResultListGameRespDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
        data:
          type: array
          items:
            $ref: '#/components/schemas/GameRespDto'
          description: List of games
        message:
          type: string
          description: Response message
    ResultGameRespDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
        data:
          $ref: '#/components/schemas/GameRespDto'
          description: Game data
        message:
          type: string
          description: Response message
    ResultString:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
        data:
          type: string
          description: Response data
        message:
          type: string
          description: Response message
    GameReqDto:
      type: object
      required:
        - name
        - title
      properties:
        name:
          type: string
          description: Unique name of the game
        title:
          type: string
          description: Title of the game
        description:
          type: string
          description: Description of the game
        instruction:
          type: string
          description: Instructions for the game
        images:
          type: array
          description: List of images associated with the game
          items:
            type: object
            allOf:
              - $ref: '#/components/schemas/ImageCreateReqDto'
        items:
          type: array
          description: List of game items
          items:
            type: object
            allOf:
              - $ref: '#/components/schemas/GameItemCreateReqDto'
    GameUpdateReqDto:
      type: object
      properties:
        title:
          type: string
          description: Title of the game
        description:
          type: string
          description: Description of the game
        instruction:
          type: string
          description: Instructions for the game
    GameRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the game
        name:
          type: string
          description: Unique name of the game
        title:
          type: string
          description: Title of the game
        description:
          type: string
          description: Description of the game
        instruction:
          type: string
          description: Instructions for the game
        images:
          type: array
          description: List of images associated with the game
          items:
            type: object
            allOf:
              - $ref: '#/components/schemas/ImageRespDto'
        items:
          type: array
          description: List of game items
          items:
            type: object
            allOf:
              - $ref: '#/components/schemas/GameItemRespDto'
    ImageCreateReqDto:
      type: object
      required:
        - imageUrl
        - gameName
      properties:
        imageUrl:
          type: string
          description: URL of the image
        gameName:
          type: string
          description: Name of the game to associate the image with
    ImageUpdateReqDto:
      type: object
      required:
        - imageUrl
      properties:
        imageUrl:
          type: string
          description: URL of the image
    ImageRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the image
        imageUrl:
          type: string
          description: URL of the image
    ResultImageRespDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
        data:
          $ref: '#/components/schemas/ImageRespDto'
          description: Image data
        message:
          type: string
          description: Response message
    GameItemCreateReqDto:
      type: object
      required:
        - name
        - type
        - region
        - price
        - gameName
      properties:
        name:
          type: string
          description: Name of the game item
        type:
          type: string
          description: Type of the game item
        region:
          type: string
          description: Region of the game item
        details:
          type: string
          description: Additional details about the game item
        price:
          type: integer
          format: int32
          description: Price of the game item
        gameName:
          type: string
          description: Name of the game to associate the item with
    GameItemUpdateReqDto:
      type: object
      required:
        - name
        - type
        - region
        - price
      properties:
        name:
          type: string
          description: Name of the game item
        type:
          type: string
          description: Type of the game item
        region:
          type: string
          description: Region of the game item
        details:
          type: string
          description: Additional details about the game item
        price:
          type: integer
          format: int32
          description: Price of the game item
    GameItemRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the game item
        name:
          type: string
          description: Name of the game item
        type:
          type: string
          description: Type of the game item
        region:
          type: string
          description: Region of the game item
        details:
          type: string
          description: Additional details about the game item
        price:
          type: integer
          format: int32
          description: Price of the game item
    ResultGameItemRespDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
        data:
          $ref: '#/components/schemas/GameItemRespDto'
          description: Game item data
        message:
          type: string
          description: Response message
    MessageToChatReq:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The message content to be sent to chat 